requirements.txt
ollama
streamlit
sentence-transformers
faiss-cpu
